Reimplement team management
The private class Team has been added to hold data about a team.
It contains just their name, score and a list of team member names.
Note there are two data structures which hold references to each team.
First, there is a map between the username of a player and the team they belong to.
Second, there is a map between the name of a team and the team itself.